push

+++ Assets/Scripts/PlayerScripts/Player.cs
@@ -29,7 +29,7 @@ public class Player : MonoBehaviour
     [SerializeField] int startingMoney;
     private float currentMana;
     private int currentHealth;
-    private int money;
+    private float money;
 
     // Start is called before the first frame update
     void Start()
@@ -59,14 +59,14 @@ public class Player : MonoBehaviour
         }
     }
 
-    public void GiveMoney(int amount)
+    public void GiveMoney(float amount)
     {
         money += amount;
     }
 
     public int GetMoney()
     {
-        return money;
+        return (int) money;
     }
 
     public void RemoveMoney(int amount)
